finalizar refatoração do processamento de dados
